finalproject/backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
letterboxdpy>=0.3.3
